アソシエーション追加、showV修正、WorksController修正

+++ db/migrate/20170413213740_create_works.rb
@@ -4,7 +4,7 @@ class CreateWorks < ActiveRecord::Migration
       t.integer :user_id
       t.date :date
       t.string :work
-      t.integer :work_time
+      t.float :work_time
       t.integer :task_id
 
       t.timestamps null: false
